jwt implementation

+++ srv/routes/user.js
@@ -1,25 +1,15 @@
 const router = require('express').Router();
 let User = require('../../db/models/user.model').User;
 
-// Get all products
+// Get all users
 router.route('/').get((req, res) => {
 	User.find()
 		.then((user) => res.json(user))
 		.catch((err) => res.status(400).json('Error: ' + err));
 });
 
-// Login authenticate
-router.route('/login').get((req, res) => {
-	const username = req.body.username;
-	const password = req.body.password;
-	User.getAuthenticated(username, password, (err, user, reason) => {
-		if (err) res.status(404).json(`${reason} : ${err}`);
-		res.status(200).json(user);
-	});
-});
-
 // Get specific user
-router.route('/:id').get((req, res) => {
+router.get('/:id', (req, res) => {
 	const id = req.params.id;
 	User.findById(id, (err, user) => {
 		if (err) res.status(400).json('Error: ' + err);
@@ -28,16 +18,25 @@ router.route('/:id').get((req, res) => {
 });
 
 // Create new user
-router.route('/').post((req, res) => {
-	const newUser = new User(req.body);
-	newUser
-		.save()
-		.then(() => res.json('User added.'))
-		.catch((err) => res.status(400).json('Error: ' + err));
+router.post('/', async (req, res) => {
+	try {
+		const userExists = await User.find({ username: req.body.username });
+		if (userExists.length) {
+			res.status(400).json('User already exists with this username');
+			return;
+		}
+		const newUser = new User(req.body);
+		newUser
+			.save()
+			.then(() => res.json('User added.'))
+			.catch((err) => res.status(400).json('Error: ' + err));
+	} catch (err) {
+		res.status(400).json('Error: ' + err);
+	}
 });
 
 // Update a specific user
-router.route('/:id').put(async (req, res) => {
+router.put('/:id', async (req, res) => {
 	const id = req.params.id;
 	try {
 		let updatedUser = await User.findByIdAndUpdate(id, req.body, {
@@ -51,13 +50,20 @@ router.route('/:id').put(async (req, res) => {
 });
 
 // Delete a user
-router.route('/:id').delete(async (req, res) => {
-	const id = req.params.id;
-	try {
-		const deletedUser = await User.findByIdAndDelete(id);
-		res.json(deletedUser);
-	} catch (err) {
-		res.status(400).json('Error: ' + err);
+router.delete('/:id', async (req, res, next) => {
+	if (req.body.role === 'ADMIN') {
+		const id = req.params.id;
+		try {
+			const deletedUser = await User.findByIdAndDelete(id);
+			res.json(deletedUser);
+		} catch (err) {
+			if (err.status) {
+				next(err);
+			}
+			res.status(400).json('Error: ' + err);
+		}
+	} else {
+		res.status(404).json(new Error('Not Authorized'));
 	}
 });
